Create ZwaveThermostatPlus
This DTH is a clone of the stock Z-wave thermostat with a few UX mods
As I did not care for the slow 4-state thermostat mode-change button, created 4 small buttons for Off, Cooling, Heating and Auto mode directly. I have not done the same for FanMode as I rarely use it, if ever. However this example should show how to easily doctor the stock DTH's to your own liking.